add basic logging for auth
